Move config code to a separate file
