+++ trillian/ctfe/instance.go
@@ -19,11 +19,9 @@ import (
 	"crypto"
 	"errors"
 	"fmt"
-	"io/ioutil"
 	"net/http"
 	"time"
 
-	"github.com/golang/protobuf/proto"
 	"github.com/golang/protobuf/ptypes"
 	"github.com/google/certificate-transparency-go/trillian/ctfe/configpb"
 	"github.com/google/certificate-transparency-go/trillian/util"
@@ -32,139 +30,6 @@ import (
 	"github.com/google/trillian/crypto/keys"
 	"github.com/google/trillian/monitoring"
 )
-
-// LogConfigFromFile creates a slice of LogConfig options from the given
-// filename, which should contain text-protobuf encoded configuration data.
-func LogConfigFromFile(filename string) ([]*configpb.LogConfig, error) {
-	cfgText, err := ioutil.ReadFile(filename)
-	if err != nil {
-		return nil, err
-	}
-
-	var cfg configpb.LogConfigSet
-	if err := proto.UnmarshalText(string(cfgText), &cfg); err != nil {
-		return nil, fmt.Errorf("failed to parse log config: %v", err)
-	}
-
-	if len(cfg.Config) == 0 {
-		return nil, errors.New("empty log config found")
-	}
-	return cfg.Config, nil
-}
-
-// ToMultiLogConfig creates a multi backend config proto from the data
-// loaded from a single-backend configuration file. All the log configs
-// reference a default backend spec as provided.
-func ToMultiLogConfig(cfg []*configpb.LogConfig, beSpec string) *configpb.LogMultiConfig {
-	defaultBackend := &configpb.LogBackend{Name: "default", BackendSpec: beSpec}
-	for _, c := range cfg {
-		c.LogBackendName = defaultBackend.Name
-	}
-	return &configpb.LogMultiConfig{
-		LogConfigs: &configpb.LogConfigSet{Config: cfg},
-		Backends:   &configpb.LogBackendSet{Backend: []*configpb.LogBackend{defaultBackend}},
-	}
-}
-
-// MultiLogConfigFromFile creates a LogMultiConfig proto from the given
-// filename, which should contain text-protobuf encoded configuration data.
-// Does not do full validation of the config but checks that it is non empty.
-func MultiLogConfigFromFile(filename string) (*configpb.LogMultiConfig, error) {
-	cfgText, err := ioutil.ReadFile(filename)
-	if err != nil {
-		return nil, err
-	}
-
-	var cfg configpb.LogMultiConfig
-	if err := proto.UnmarshalText(string(cfgText), &cfg); err != nil {
-		return nil, fmt.Errorf("failed to parse multi-backend log config: %v", err)
-	}
-
-	if len(cfg.LogConfigs.GetConfig()) == 0 || len(cfg.Backends.GetBackend()) == 0 {
-		return nil, errors.New("config is missing backends and/or log configs")
-	}
-	return &cfg, nil
-}
-
-// ValidateLogMultiConfig checks that a config is valid for use with multiple
-// backend log servers. The rules applied are:
-//
-// 1. The backend set must define a set of log backends with distinct
-// (non empty) names and non empty backend specs.
-// 2. The backend specs must all be distinct.
-// 3. The log configs must all specify a log backend and each must be one of
-// those defined in the backend set.
-// 4. If NotBeforeStart or NotBeforeLimit are set for a log then these fields
-// must be valid timestamp protos. If both are set then NotBeforeLimit must
-// not be before NotBeforeStart.
-// 5. The prefixes of configured logs must all be distinct and must not be
-// empty.
-// 6. The set of tree ids for each configured backend must be distinct.
-func ValidateLogMultiConfig(cfg *configpb.LogMultiConfig) (map[string]*configpb.LogBackend, error) {
-	// Check the backends have unique non empty names and build the map.
-	backendMap := make(map[string]*configpb.LogBackend)
-	bSpecMap := make(map[string]bool)
-	for _, backend := range cfg.Backends.Backend {
-		if len(backend.Name) == 0 {
-			return nil, fmt.Errorf("empty backend name: %v", backend)
-		}
-		if len(backend.BackendSpec) == 0 {
-			return nil, fmt.Errorf("empty backend_spec for backend: %v", backend)
-		}
-		if _, ok := backendMap[backend.Name]; ok {
-			return nil, fmt.Errorf("duplicate backend name: %v", backend)
-		}
-		if ok := bSpecMap[backend.BackendSpec]; ok {
-			return nil, fmt.Errorf("duplicate backend spec: %v", backend)
-		}
-		backendMap[backend.Name] = backend
-		bSpecMap[backend.BackendSpec] = true
-	}
-
-	// Check that logs all reference a defined backend and there are no duplicate
-	// or empty prefixes. Apply other LogConfig specific checks.
-	logNameMap := make(map[string]bool)
-	logIDMap := make(map[string]bool)
-	for _, logCfg := range cfg.LogConfigs.Config {
-		if len(logCfg.Prefix) == 0 {
-			return nil, fmt.Errorf("log config: empty prefix: %v", logCfg)
-		}
-		if logNameMap[logCfg.Prefix] {
-			return nil, fmt.Errorf("log config: duplicate prefix: %s: %v", logCfg.Prefix, logCfg)
-		}
-		if _, ok := backendMap[logCfg.LogBackendName]; !ok {
-			return nil, fmt.Errorf("log config: references undefined backend: %s: %v", logCfg.LogBackendName, logCfg)
-		}
-		logNameMap[logCfg.Prefix] = true
-		logIDKey := fmt.Sprintf("%s-%d", logCfg.LogBackendName, logCfg.LogId)
-		if ok := logIDMap[logIDKey]; ok {
-			return nil, fmt.Errorf("log config: dup tree id: %d for: %v", logCfg.LogId, logCfg)
-		}
-		var err error
-		var tStart time.Time
-		start := logCfg.GetNotAfterStart()
-		if start != nil {
-			tStart, err = ptypes.Timestamp(start)
-			if err != nil {
-				return nil, fmt.Errorf("log_config: invalid start timestamp %v for: %v", err, logCfg)
-			}
-		}
-		var tLimit time.Time
-		limit := logCfg.GetNotAfterLimit()
-		if limit != nil {
-			tLimit, err = ptypes.Timestamp(limit)
-			if err != nil {
-				return nil, fmt.Errorf("log_config: invalid limit timestamp %v for: %v", err, logCfg)
-			}
-		}
-		if start != nil && limit != nil && tLimit.Before(tStart) {
-			return nil, fmt.Errorf("log_config: limit before start for: %v", logCfg)
-		}
-		logIDMap[logIDKey] = true
-	}
-
-	return backendMap, nil
-}
 
 var stringToKeyUsage = map[string]x509.ExtKeyUsage{
 	"Any":                        x509.ExtKeyUsageAny,
